components/flight-map: Import "map-hover-handler" code
