+++ ember/app/components/flight-map.js
@@ -2,34 +2,65 @@ import Ember from 'ember';
 
 import BaseMapComponent from './base-map';
 import slMapClickHandler from '../utils/map-click-handler';
-import slMapHoverHandler from '../utils/map-hover-handler';
 
 export default BaseMapComponent.extend({
   fixCalc: Ember.inject.service(),
 
   flights: Ember.computed.readOnly('fixCalc.flights'),
 
+  hoverEnabled: true,
+
   didInsertElement() {
     this._super(...arguments);
-    this.get('map').on('moveend', this._handleMoveEnd, this);
+
+    let map = this.get('map');
+    map.on('moveend', this._handleMoveEnd, this);
+    map.on('pointermove', this._handlePointerMove, this);
 
     let fixCalc = this.get('fixCalc');
-
-    slMapHoverHandler.create({
-      fixCalc,
-      flightMap: this,
-    });
 
     slMapClickHandler(this.get('map'), fixCalc.get('flights'));
   },
 
   willDestroyElement() {
     this._super(...arguments);
-    this.get('map').off('moveend', this._handleMoveEnd, this);
+    let map = this.get('map');
+    map.off('moveend', this._handleMoveEnd, this);
+    map.off('pointermove', this._handlePointerMove, this);
   },
 
   _handleMoveEnd(event) {
     this.getWithDefault('onExtentChange', Ember.K)(event.frameState.extent);
+  },
+
+  _handlePointerMove(event) {
+    if (event.dragging || !this.get('hoverEnabled')) return;
+
+    let map = this.get('map');
+    let source = this.get('flights.source');
+
+    let coordinate = map.getEventCoordinate(event.originalEvent);
+    let feature = source.getClosestFeatureToCoordinate(coordinate);
+
+    if (feature !== null) {
+      let geometry = feature.getGeometry();
+      let closest_point = geometry.getClosestPoint(coordinate);
+
+      let feature_pixel = map.getPixelFromCoordinate(closest_point);
+      let mouse_pixel = map.getPixelFromCoordinate(coordinate);
+
+      let squared_distance = Math.pow(mouse_pixel[0] - feature_pixel[0], 2) +
+        Math.pow(mouse_pixel[1] - feature_pixel[1], 2);
+
+      // Set the time when the mouse hovers the map
+      if (squared_distance > 100) {
+        this.get('fixCalc').resetTime();
+      } else {
+        this.get('fixCalc').set('time', closest_point[3]);
+      }
+
+      map.render();
+    }
   },
 
   actions: {
