+++ ember/app/templates/components/tracking-page.hbs
@@ -19,6 +19,7 @@
 </div>
 
 {{flight-map
+  hoverEnabled=(not (or fixCalc.isRunning cesiumEnabled))
   onExtentChange=(action (mut mapExtent))
   onCesiumEnabledChange=(action (mut cesiumEnabled))
   fullscreenElement="#fullscreen-content"
